src: dri2: fix MSC handling when a pixmap is moved between CRTCs
We need to ensure that a drawable has monotonic MSC values irrespective
of what happens to the drawable - this includes when a drawable is
dragged between different CRTCs.  Arrange to handle this, by noticing
that the CRTC has changed, and adapt the MSC value accordingly.  We
also need to handle this when setting an event against the current
CRTC.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>