src: dri2: fix MSC handling
The MSC handling was taken from the xf86-video-intel driver, but it
seems to be lacking in that the wrap-handling is substandard; it can
result in DRI2 applications stalling for a very long time.

The problem is that xf86-video-intel is under the impression that the
vblank counter can wind backwards.  This is not true; the kernel only
ever increments the counter (see functions that call store_vblank().)
Hence, remove this logic, and always assume that a vblank sequence
value that is less than a previously seen one means that it has
wrapped.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>