Fix segfault in glyph cache
When the X server unrealises a glyph, we weren't removing it from the
glyph cache.  This results in a stale pointer in the cache to the glyph
which then is dereferenced when looking for a glyph to evict, leading
eventually to a segfault.

Fix this by removing the glyph from the cache while it's being
unrealised.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>