etnaviv: fall back if composite region is bigger than src or mask
Having a composite region bigger than the src or mask picture is well
defined as per the X.Org renderproto spec chapter
"9. Source and Mask Transformations".

We can not do this on the GPU as the core is not clipping the source
size in relative mode and there is no way to control the behavior in
absolute mode.

This is especially visible on a GPU with MMUv2, where such operations
will cause MMU faults, as the GPU tries to read outside of the source
buffer. On MMUv1 it will only cause reading of undefined data.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Russell King <rmk@armlinux.org.uk>