src: Xv: add support for disabling the primary plane
When Xv video is being displayed at full screen, and the video is
completely unobscured, there is no point having the primary plane
enabled; this wastes system bus bandwidth as the display hardware has
to read the primary plane to check the colorkey.

Disable the primary plane in this case, restoring it when the primary
plane becomes visible once again.

This results in a gain of CPU performance of about 20% when video is
displayed at full screen on the Dove Cubox.

Signed-off-by: Russell King <rmk@armlinux.org.uk>