etnaviv: avoid out of bounds accelerated picture accesses
The GPU can't handle repeats or accesses outside of the source pictures
so we need to be careful to handle these correctly.  If the source
coordinates fall entirely within the source picture, we're fine.

Replace picture_needs_repeat() with picture_has_pixels() which validates
that the region we're going to fetch is covered by the underlying
drawable.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>