src: bump minimum libdrm to 2.4.47 and use drmSetClientCap()
drmSetClientCap() was added in libdrm version 2.4.47, which is a useful
wrapper for our universal plane support.  Use this instead of open
coding this functionality.

DRM_CLIENT_CAP_UNIVERSAL_PLANES was introduced in 2.4.56, but we don't
want to force that version just for this definition.  Provide a compat
definition for this.

Signed-off-by: Russell King <rmk@armlinux.org.uk>